Implementing bonus checker part

+++ src/debug.c
@@ -12,7 +12,7 @@
 
 #include "../include/push_swap.h"
 
-int	is_sorted(t_stacks *s)
+uint8_t	is_sorted(t_stacks *s)
 {
 	int	i;
